Subtitles: Search by URL not assumed index
Currently, when populating subtitleTracks, we assume that the ordering
and list of populated subtitle tracks will not change when Roku moves
the list into availableSubtitleTracks. This causes an issue with some
videos as it is not always consistent.

This patch modifies the logic to no-longer inject assumed final indices
into our list of text-based subtitles, but instead search through the
availableSubtitleTracks array and locate the actual subtitle that
refers to the same URL as in our list. In this way we are guaranteed
to always tell Roku to play the subtitle we want, no matter how re-
ordered the options get.

NOTE: The URL gets mildly mangled in the process of copying from
subtitleTracks to availableSubtitleTracks, so we need so search via
substring, rather than doing a full string comparison.